pyproject.toml
[project]
name = "model-serving-platform"
version = "0.1.0"
description = "OCR model serving platform"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "gunicorn>=21.2",
  "prometheus-client>=0.20",
  "opentelemetry-api>=1.23",
  "opentelemetry-sdk>=1.23",
  "opentelemetry-instrumentation-fastapi>=0.44b0",
  "SQLAlchemy>=2.0",
  "aiomysql>=0.2.0",
  "cryptography>=42.0",
  "greenlet>=3.0",
  "python-dotenv>=1.0",
  "redis>=5.0",
  "kafka-python>=2.0",
  "aiokafka>=0.10.0",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "pydantic-core>=2.0",
  "pydantic-settings>=2.0",

]

[dependency-groups]
dev = ["pytest", "pytest-asyncio", "httpx", "ruff"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["app"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "black",
  "mypy",
]

obs = [
  "opentelemetry-exporter-otlp",
  "opentelemetry-instrumentation-logging",
]
